Tag Service tests moved to folder Tag
