+++ tests/Feature/Service/Student/StudentProjectsDetailServiceTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Service\Student;
+namespace Tests\Feature\Service\Student\Tag;
 
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
